Added new scene for begining of each save
Can copy and paste the canvas to other scene in order to add cut scenes and dialog with other players that change on click. add phrases to the list by clicking on the canvas and looking for the introductiontext script on the right side and add phrases to the array inside the editor.